+++ src/OneOffGamesPlugin/OneOffGamesPlugin.fs
@@ -47,6 +47,7 @@ type OneOffGamesPlugin() =
             ("IHF Handball Challenge 12", SimpleGames.ExtractIHFHandballChallenge12)
             ("IHF Handball Challenge 14", SimpleGames.ExtractIHFHandballChallenge14)
             ("Torment: Tides of Numenera", PillarsOfEternity.ExtractTormentTidesOfNumenera)
+            ("Tyranny", PillarsOfEternity.ExtractTyranny)
         |] |> Map.ofArray
 
     interface IPlugin with