+++ RELEASE_NOTES.md
@@ -1,5 +1,8 @@
 #### 0.14.0 - July 5 2017
 * Added command line options for 'export-anki' to limit the length (in characters or whitespace-delimited words) of the text of a card: --recognition-length-limit, --production-length-limit, --recognition-word-limit, and --production-word-limit.
+* Added support for:
+    * [Torment: Tides of Numenera](http://store.steampowered.com/app/272270/Torment_Tides_of_Numenera/) \[de, en, es, fr, it, pl, ru\]
+	* [Tyranny](http://store.steampowered.com/app/362960/Tyranny/) \[de, en, es, fr, pl, ru\]
 
 #### 0.13.0 - November 29 2016
 * Added support for: